execute stage1 install scripts on pkg installation

+++ lpm/core/src/delete.rs
@@ -42,7 +42,9 @@ impl PkgDeleteTasks for PkgDataFromDb {
         }
 
         let pkg_lib_dir = Path::new("/var/lib/lpm/pkg").join(&self.meta_dir.meta.name);
-        fs::remove_dir_all(pkg_lib_dir)?;
+        if Path::new(&pkg_lib_dir).exists() {
+            fs::remove_dir_all(pkg_lib_dir)?;
+        }
 
         transaction_op(&db, Transaction::Commit)?;
         db.close();
